+++ langchain4j-dashscope/src/main/java/dev/langchain4j/model/dashscope/QwenStreamingLanguageModel.java
@@ -7,7 +7,7 @@ public class QwenStreamingLanguageModel extends QwenStreamingChatModel implement
     protected QwenStreamingLanguageModel(String apiKey,
                                          String modelName,
                                          Double topP,
-                                         Double topK,
+                                         Integer topK,
                                          Boolean enableSearch,
                                          Integer seed) {
         super(apiKey, modelName, topP, topK, enableSearch, seed);
@@ -35,7 +35,7 @@ public class QwenStreamingLanguageModel extends QwenStreamingChatModel implement
             return (Builder) super.topP(topP);
         }
 
-        public Builder topK(Double topK) {
+        public Builder topK(Integer topK) {
             return (Builder) super.topK(topK);
         }
